event_backend/requirements.txt
psycopg2-binary==2.9.10
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.5.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
dj-database-url==3.0.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
msgpack==1.1.1
redis==6.2.0
daphne>=4.0.0

